requirements.txt
Django~=5.2.4
social-auth-core~=4.7.0
pandas~=2.3.1
djangorestframework~=3.16.0